requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
argcomplete==3.6.2
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
boto3==1.37.30
botocore==1.37.30
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.8
cohere==5.14.2
colorama==0.4.6
coloredlogs==15.0.1
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.11
fastavro==1.10.0
filelock==3.18.0
Flask==2.3.3
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-WTF==1.1.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gevent==24.11.1
geventhttpclient==2.3.3
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
greenlet==3.2.2
griffe==1.7.2
groq==0.22.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
kiwisolver==1.4.8
locust==2.37.1
locust-cloud==1.21.3
logfire-api==3.12.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mcp==1.6.0
mcp-server-time==0.6.2
mdurl==0.1.2
mistralai==1.6.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
numpy==2.2.4
onnxruntime==1.21.0
openai==1.68.0
opentelemetry-api==1.31.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
peewee==3.16.3
pillow==11.1.0
platformdirs==4.3.8
plotly==5.17.0
posthog==3.21.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.30.1
psutil==7.0.0
psycopg2-binary==2.9.7
pulsar-client==3.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-ai==0.0.53
pydantic-ai-slim==0.0.53
pydantic-evals==0.0.53
pydantic-graph==0.0.53
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PyPika==0.48.9
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
sse-starlette==2.2.1
starlette==0.46.1
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
Wikipedia-API==0.8.1
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
